[refactor] Migrate Product, ProductItem from inside of Homepage folder to components folder.
